refactor: update aggregate builder and introduce mapping functions
- Removed unused imports and refactored type definitions in aggregate-builder.ts
- Introduced new mapping functions for command and event handling in separate files
- Enhanced the createEventDeciderFn and createReducerFn functions to use the new mapping functions
- Improved type safety and clarity in the aggregate builder implementation
